# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipper",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/accounts/{account_id}": {
      "patch": {
        "summary": "Update Account",
        "operationId": "update_account",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/next-slot": {
      "get": {
        "summary": "Next Slot",
        "operationId": "next_slot",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextSlot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "summary": "List Posts",
        "description": "Sorted by scheduled_for. from/to filter on coalesce(published_at, scheduled_for).",
        "operationId": "list_posts",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "brand_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "DRAFT",
                      "SCHEDULED",
                      "PUBLISHING",
                      "PUBLISHED",
                      "FAILED",
                      "DEAD_LETTER",
                      "CANCELLED"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostOut"
                  },
                  "title": "Response List Posts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Post",
        "operationId": "create_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}": {
      "get": {
        "summary": "Get Post",
        "operationId": "get_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Post",
        "operationId": "update_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/auto-schedule": {
      "post": {
        "summary": "Auto Schedule",
        "operationId": "auto_schedule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoScheduleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoScheduleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/approve": {
      "post": {
        "summary": "Approve Post",
        "operationId": "approve_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/cancel": {
      "post": {
        "summary": "Cancel Post",
        "operationId": "cancel_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "zernio_account_id": {
            "type": "string",
            "title": "Zernio Account Id"
          },
          "zernio_profile_id": {
            "type": "string",
            "title": "Zernio Profile Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "connection_status": {
            "type": "string",
            "enum": [
              "connected",
              "disconnected"
            ],
            "title": "Connection Status"
          },
          "posting_slots": {
            "$ref": "#/components/schemas/PostingSlots"
          },
          "daily_cap": {
            "type": "integer",
            "title": "Daily Cap"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "min_gap_minutes": {
            "type": "integer",
            "title": "Min Gap Minutes"
          },
          "connected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Connected At"
          },
          "last_publish_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Publish At"
          },
          "disabled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "today_count": {
            "type": "integer",
            "title": "Today Count",
            "default": 0
          },
          "quota": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quota"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_post_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Post At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "zernio_account_id",
          "zernio_profile_id",
          "username",
          "avatar_url",
          "connection_status",
          "posting_slots",
          "daily_cap",
          "timezone",
          "min_gap_minutes",
          "connected_at",
          "last_publish_at",
          "disabled_at"
        ],
        "title": "AccountOut"
      },
      "AccountPatch": {
        "properties": {
          "posting_slots": {
            "$ref": "#/components/schemas/PostingSlots"
          },
          "daily_cap": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Daily Cap"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "min_gap_minutes": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 0.0,
            "title": "Min Gap Minutes"
          },
          "disabled": {
            "type": "boolean",
            "title": "Disabled"
          }
        },
        "type": "object",
        "title": "AccountPatch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NextSlot": {
        "properties": {
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          }
        },
        "type": "object",
        "required": [
          "scheduled_for"
        ],
        "title": "NextSlot"
      },
...
      "PostCreate": {
        "properties": {
          "render_id": {
            "type": "integer",
            "title": "Render Id"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "scheduled_for": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled For"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2200
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "render_id",
          "account_id",
          "scheduled_for"
        ],
        "title": "PostCreate"
      },
...
      "PostPatch": {
        "properties": {
          "scheduled_for": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled For"
          },
          "caption": {
            "type": "string",
            "maxLength": 2200,
            "title": "Caption"
          }
        },
        "type": "object",
        "title": "PostPatch"
      },
...
    }
  }
}